requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
redis
